Added some interfaces for tourists (user profile and reservations)

+++ resources/views/user/edit.blade.php
@@ -93,12 +93,14 @@
                             {!! Form::submit('Edit User', ['class'=>'btn btn-primary']) !!}
                         </div>
                         {!! Form::close() !!}
-                        <div class="form-group">
-                            {!!  Form::open(['method' => 'DELETE', 'action' => ['UserController@destroy', $user->id]])!!}
-                            {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                            {!! Form::close() !!}
-                            <button type="button" class="btn btn-primary" onclick="location.href='{{route('user.index')}}'">Cancel</button>
-                        </div>
+                        @if(Auth::user()->user_role_id == 3)
+                            <div class="form-group">
+                                {!!  Form::open(['method' => 'DELETE', 'action' => ['UserController@destroy', $user->id]])!!}
+                                {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                                {!! Form::close() !!}
+                            </div>
+                        @endif
+                        <button type="button" class="btn btn-primary" onclick="location.href='{{route('user.index')}}'">Cancel</button>
                     </div>
                 </div>
             </div>
